se añaden enums para Status y Rating, se crea la clase abstracta UserMediaContent y sus clases hijas con idicaciones para futuras implementaciones de los metodos abstractos.

+++ src/VideoURL.java
@@ -1,20 +1,23 @@
-
 public class VideoURL {
 	
-	private int videoId;
+	//Atributos de clase
+	private Long videoId;
 	private String urlList;	
 
+	//Constructores
 	public VideoURL () {
 		
 	}
 	
-	public VideoURL (int videoId, String urlList) {
+	public VideoURL (Long videoId, String urlList) {
 		
 		this.videoId = videoId;
 		this.urlList = urlList;
 		
 	}
+	
 
+	//Setters y Getters
 	public String getURL_list() {
 		return urlList;
 	}
@@ -23,7 +26,7 @@ public class VideoURL {
 		this.urlList = urlList;
 	}
 
-	public int getVideo_Id() {
+	public Long getVideo_Id() {
 		return videoId;
 	}
